+++ specs/Data/MediaBus/Conduit/SegmentSpec.hs
@@ -5,87 +5,70 @@ import Data.MediaBus
 import Data.MediaBus.Conduit.SpecUtils
 import Test.Hspec
 import Test.QuickCheck
-import Data.Word (Word32)
 
 spec :: Spec
 spec = do
-  it "sum of the durations of the generated segments is equal to the sum of the input frames rounded to the next multiple of the target duration"
-    $ property
-    $ \ls ->
-      let
-        outputPayload = PT1024
-        outputSegmentDuration :: Ticks (Hz 16000) Word32
-        outputSegmentDuration = getDurationTicks outputPayload
-      in runSegmetCOnTestData @(Hz 16000)
-        ls
-        outputPayload
-        ( \inputs outputs ->
-            let inputSum = sum (map getDurationTicks inputs)
-                outputSum = sum (map getDurationTicks outputs)
-            in ((inputSum `div` outputSegmentDuration) * outputSegmentDuration)
-                  === outputSum .||.
-                    inputSum === outputSum -- for debugging
-        )
-  it "if the output length is zero, no output is generated"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData  @(Hz 8000)
-        ls
-        PT0
-        ( \inputs outputs -> not (null inputs) ==> outputs === [] )
-  it "only outputs segments with a valid length"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData  @(Hz 8000)
-        ls
-        PT5
-        ( \_ outputs ->
-            Prelude.drop 1 outputs
-              `shouldSatisfy` all ((== getDuration PT5) . getDuration)
-        )
-  it "only outputs segments with a valid length"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData  @(Hz 8000)
-        ls
-        PT20
-        ( \_ outputs ->
-            Prelude.drop 1 outputs
-              `shouldSatisfy` all ((== getDuration PT20) . getDuration)
-        )
-  it "only outputs segments with a valid length"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData  @(Hz 8000)
-        ls
-        PT1024
-        ( \_ outputs ->
-            Prelude.drop 1 outputs
-              `shouldSatisfy` all ((== getDuration PT1024) . getDuration)
-        )
-  it "only drops less than the static duration"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData @(Hz 8000)
-        ls
-        PT5
-        ( \inputs outputs ->
-            ( sum (getDuration <$> inputs)
-                - sum (getDuration <$> outputs)
-            )
-              `shouldSatisfy` (< getDuration PT5)
-        )
-  it "maintains strictly monotonic sequence numbers"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData @(Hz 8000)
-        ls
-        PT5
-        (const (`shouldSatisfy` seqNumStrictlyMonotoneIncreasing))
-  it "maintains strictly monotonic timestamps"
-    $ property
-    $ \ls ->
-      runSegmetCOnTestData @(Hz 8000)
-        ls
-        PT5
-        (const (`shouldSatisfy` ticksStrictlyMonotoneIncreasing (nominalDiffTime # getStaticDuration PT5)))
+  it "generates segments with a total duration equal to the input frames" $
+    property $
+      \ls (outputSegmentDuration :: Ticks64 (Hz 16000)) ->
+        runSegmetCOnTestData @(Hz 16000)
+          ls
+          (outputSegmentDuration ^. nominalDiffTime)
+          ( \inputs outputs ->
+              let inputSum, outputSum :: Ticks64 (Hz 16000)
+                  inputSum = sum (map getDurationTicks inputs)
+                  outputSum = sum (map getDurationTicks outputs)
+               in outputSum === inputSum
+          )
+  it "generates incomplete frames with previous left over input when it receives a start frame" $
+    property $
+      \ls1 ls2 (outputSegmentDuration :: Ticks64 (Hz 16000)) ->
+        not (null ls1) && not (null ls2) && outputSegmentDuration > 0
+          ==> let inputs = mkTestInputs @(Hz 16000) ls1 ++ mkTestInputs @(Hz 16000) ls2
+                  outputs = runSegmetC inputs (outputSegmentDuration ^. nominalDiffTime)
+                  inputSum, outputSum :: Ticks64 (Hz 16000)
+                  inputSum = sum (map getDurationTicks inputs)
+                  outputSum = sum (map getDurationTicks outputs)
+               in outputSum === inputSum
+  it "starts with the sequence numbers of the first frame" $
+    property $
+      \ls1 ls2 (outputSegmentDuration :: Ticks64 (Hz 16000)) ->
+        not (null ls1) && not (null ls2) && outputSegmentDuration > 0
+          ==> let inputs = mkTestInputs @(Hz 16000) ls1 ++ mkTestInputs @(Hz 16000) ls2
+                  outputs = runSegmetC inputs (outputSegmentDuration ^. nominalDiffTime)
+                  inputSum, outputSum :: Ticks64 (Hz 16000)
+                  inputSum = sum (map getDurationTicks inputs)
+                  outputSum = sum (map getDurationTicks outputs)
+               in outputSum === inputSum
+
+  it "generates no output if the output length is zero" $
+    property $
+      \ls ->
+        runSegmetCOnTestData @(Hz 8000)
+          ls
+          0
+          (\inputs outputs -> not (null inputs) ==> outputs === [])
+  it "generates segments with the given suration, for the last one" $
+    property $
+      \ls ->
+        runSegmetCOnTestData @(Hz 8000)
+          ls
+          0.005
+          ( \_ outputs ->
+              Prelude.drop 1 (Prelude.take (length outputs - 2) outputs)
+                `shouldSatisfy` all ((== 0.005) . getDuration)
+          )
+  it "maintains strictly monotonic sequence numbers" $
+    property $
+      \ls ->
+        runSegmetCOnTestData @(Hz 8000)
+          ls
+          0.005
+          (const (`shouldSatisfy` seqNumStrictlyMonotoneIncreasing))
+  it "maintains strictly monotonic timestamps" $
+    property $
+      \ls ->
+        runSegmetCOnTestData @(Hz 8000)
+          ls
+          0.005
+          (const (`shouldSatisfy` ticksStrictlyMonotoneIncreasing (nominalDiffTime # getStaticDuration PT5)))